+++ app/policies/organization_policy.ts
@@ -21,6 +21,10 @@ export default class OrganizationPolicy extends BasePolicy {
     return false
   }
 
+  configureBranding(_user: User): AuthorizerResponse {
+    return false
+  }
+
   viewAuditLog(user: User): AuthorizerResponse {
     return user.organizationId !== null
   }